+++ ftw/upgrade/configure.zcml
@@ -8,6 +8,7 @@
 
     <i18n:registerTranslations directory="locales" />
     <include package=".browser" />
+    <include package=".jsonapi" />
 
     <genericsetup:registerProfile
         name="default"
